ENH: add standard codes to describe measurements
Also initialize quantity and units from the scalar volume being analyzed. Note
that this approach is error prone, since it cannot be guaranteed that the
scalar volume on the parameter node will be the same between the time
measurements are calculated, and the time when the getInfo is called. It is
also not guaranteed that scalar volume parameter is populated at the time
getInfo is called. We should probably initialize info at the time measurements
are calculated.